requirements.txt
openai
notion-client
streamlit
python-dotenv
